feature: add progress loading bar to top of page
The loading bar is disabled by default. To enable it in a project, add this to config.ts:
app.enableRouterLoadingBar: true

The loading bar is activated while the router is busy transitioning to a new page, but not while the page loads it's data.

To customize the color of the loading bar, modify this CSS variable in custom.scss:
--route-loading-progress-bar-color