fix: add support for multilingual read texts in mobile mode
Also refactored the mobile mode collection text page so that all available mobile mode view types are loaded, which means you can switch between them without reloading (i.e. resetting) them.

The views queryParams is set to the available mobile mode views, but is not otherwise used. This needs to be refactored in the future.