feature: add page not found page (404 page)
URL paths that don't match any routes will redirect to a 404 page. The message which is displayed on this page can be customized either by editing the translation with key PageNotFound.DefaultMessage, or by providing the message content in a markdown file from the backend. The md-file must be in the language root of the /md/ backend folder. The filename must begin with the identifier 404. Valid filenames are, for example: 404.md, 404 - page not found.md.