fix: ms and variant not loading on invalid type id in URL
If the id of a manuscript or a variant in the URL is invalid with regards to the text id, the first manuscript or variant is displayed instead of an infinite loading spinner. Also, if there is only one manuscript or variant, the URL is not updated with the type id.