feature: support adding Open Graph meta tags to document head
Open Graph meta tags (https://ogp.me/) can now be generated automatically to the app's document head for each page.

To enable this feature, set app.openGraphMetaTags.enabled to true in config.ts.

If enabled, the following new information needs to be added to the config:

In each object in the app.i18n.languages array, add the key "region". It's value must be a two letter uppercase region identifier. For instance, in the locale ID sv-FI, FI is the region identifier.

Add app.openGraphMetaTags.image as an object. It has to have a sub-object for each language with two keys: URL and altText. URL is the URL to an image file for the Open Graph protocol, and altText is the alt-text of the image. For example:

openGraphMetaTags: {
      enabled: true,
      image: {
        sv: {
          altText: "alt-text",
          URL: "assets/images/home-page-banner.jpg"
        },
        fi: {
          altText: "alt-teksti",
          URL: "assets/images/home-page-banner.jpg"
        }
     }
 }