feature: make info-overlays keyboard accessible
Focus is set to the info-overlay element when it is opened, and focus is returned to the triggering element when the info-overlay is closed.

This commit does not modify the HTML that triggers info-overlays to be opened. This means that the backend needs to provide HTML where tooltip/info-overlay triggering elements are keyboard focusable. Since these elements often are <span> they are not natively focusable, and thus not reachable by keyboard navigation. You need to add the tabindex="0" attribute to them for them to become reachable by keyboard navigation.