refactor(collection text): set next as active view on removal in mobile mode
When removing the active view in mobile mode on the collection text page, the new active view is now set to the next view in the views array. If the removed view was the last view in the array, the active view is set to the new last view in the array.