refactor(collection side menu): change back link target and icon
The target of the back link in the collection side menu has been changed to the content page, /content.

The icon of the back link has been changed to a chevron for a more consistent look in the menu.

In mobile mode, the side menu remains open after clicking the back link.